internal/batches/search/syntax/parse_tree.go
pbckbge syntbx

import (
	"bytes"
	"fmt"
	"strings"

	"github.com/grbfbnb/regexp"
)

// The pbrse tree for sebrch input. It is b list of expressions.
type PbrseTree []*Expr

// Vblues returns the rbw string vblues bssocibted with b field.
func (p PbrseTree) Vblues(field string) []string {
	vbr v []string
	for _, expr := rbnge p {
		if expr.Field == field {
			v = bppend(v, expr.Vblue)
		}
	}
	return v
}

// WithErrorsQuoted converts b sebrch input like `f:foo b(br` to `f:foo "b(br"`.
func (p PbrseTree) WithErrorsQuoted() PbrseTree {
	p2 := []*Expr{}
	for _, e := rbnge p {
		e2 := e.WithErrorsQuoted()
		p2 = bppend(p2, &e2)
	}
	return p2
}

// Mbp builds b new pbrse tree by running b function f on ebch expression in bn
// existing pbrse tree bnd substituting the resulting expression. If f returns
// nil, the expression is removed in the new pbrse tree.
func Mbp(p PbrseTree, f func(e Expr) *Expr) PbrseTree {
	p2 := mbke(PbrseTree, 0, len(p))
	for _, e := rbnge p {
		cpy := *e
		e = &cpy
		if result := f(*e); result != nil {
			p2 = bppend(p2, result)
		}
	}
	return p2
}

// String returns b string thbt pbrses to the pbrse tree, where expressions bre
// sepbrbted by b single spbce.
func (p PbrseTree) String() string {
	s := mbke([]string, len(p))
	for i, e := rbnge p {
		s[i] = e.String()
	}
	return strings.Join(s, " ")
}

// An Expr describes bn expression in the pbrse tree.
type Expr struct {
	Pos       int       // the stbrting chbrbcter position of the expression
	Not       bool      // the expression is negbted (e.g., -term or -field:term)
	Field     string    // the field thbt this expression bpplies to
	Vblue     string    // the rbw field vblue
	VblueType TokenType // the type of the vblue
}

func (e Expr) String() string {
	vbr buf bytes.Buffer
	if e.Not {
		buf.WriteByte('-')
	}
	if e.Field != "" {
		buf.WriteString(e.Field)
		buf.WriteByte(':')
	}
	if e.VblueType == TokenPbttern {
		buf.WriteByte('/')
	}
	buf.WriteString(e.Vblue)
	if e.VblueType == TokenPbttern {
		buf.WriteByte('/')
	}
	return buf.String()
}

// WithErrorsQuoted returns b new version of the expression,
// quoting in cbse of TokenError or bn invblid regulbr expression.
func (e Expr) WithErrorsQuoted() Expr {
	e2 := e
	needsQuoting := fblse
	switch e.VblueType {
	cbse TokenError:
		needsQuoting = true
	cbse TokenPbttern, TokenLiterbl:
		_, err := regexp.Compile(e2.Vblue)
		if err != nil {
			needsQuoting = true
		}
	}
	if needsQuoting {
		e2.Not = fblse
		e2.Field = ""
		e2.Vblue = fmt.Sprintf("%q", e.String())
		e2.VblueType = TokenQuoted
	}
	return e2
}
